Added lab 9 files. Updated Readme.md

+++ src/main/java/ua/lviv/iot/stationery/models/WritingItem.java
@@ -17,8 +17,8 @@ public class WritingItem extends Item {
     private String inkType;
     private WritingAids writingAid;
 
-    public WritingItem(final String name, final float weight, final float price,
-                       final float rating, final CustomerType customerType,
+    public WritingItem(final String name, final double weight, final double price,
+                       final double rating, final CustomerType customerType,
                        final String thisInkType, final WritingAids thisWritingAid) {
         super(name, weight, price, rating, customerType);
         this.inkType = thisInkType;
